Removed duplicated getStatementParent function. Refactored all babel-traverse tests to use ESmodules
