+++ packages/babel-traverse/test/ancestry.js
@@ -1,6 +1,7 @@
-const traverse = require("../lib").default;
-const assert = require("assert");
-const parse = require("babylon").parse;
+import traverse from "../lib";
+import assert from "assert";
+import { parse } from "babylon";
+import { expect } from "chai";
 
 describe("path/ancestry", function () {
   describe("isAncestor", function () {
@@ -60,6 +61,19 @@ describe("path/ancestry", function () {
       const [ , numberPath, stringPath ] = paths;
 
       assert(!numberPath.isDescendant(stringPath));
+    });
+  });
+
+  describe("getStatementParent", function () {
+    const ast = parse("var a = 1;");
+    it("should throw", function () {
+      expect(function () {
+        traverse(ast, {
+          Program(path) {
+            path.getStatementParent();
+          },
+        });
+      }).to.throw(/File\/Program node/);
     });
   });
 });
